GeometricFilter.jl

The aim of this package is to allow to perform filtering for dynamical systems on manifold, in particular homogeneous spaces.

A homogeneous space is a manifold $\mathcal{M}$ endowed with a group action from a Lie group $G$.

GeometricFilter.ActionObserver — Type

ActionObserver(A::AbstractGroupAction{<:ActionDirection}, ref)

Models the observation of a Lie group element χ as the action of χ on some manifold N.

If it is a left action, the corresponding observation is χ ↦ χ ⋅ x₀, where x₀ ∈ N is a reference point. This often corresponds to a measurement of the state, for instance, the position.

If it is a right action, the corresponding observation is $χ ↦ χ^{-1} ⋅ x₀$, where $x₀ ∈ N$ is a reference point. This often corresponds to the distance and bearing measurement of a fixed landmark.

GeometricFilter.EmptyObservation — Type

EmptyObservation()

An empty observation, that is, an observation without any useful information to update the state's uncertainty.

GeometricFilter.IdentityObserver — Type

IdentityObserver(M)

Direct observation of a point on the state manifold.

M = Sphere(2)
x = rand(M)
obs = IdentityObserver(M)
obs(x) # x

GeometricFilter.LinearObserver — Type

LinearObserver(H::AbstractMatrix,
  conv=LeftAction();
  depth=0)

Linear observation model, $x ↦ Hx$, or $x ↦ xH^*$.

H = randn(2,3)
lobs = LinearObserver(H) # standard linear observer for vectors of size 3
x = randn(3)
lobs(x) == H*x # true

A positive depth parameter indicates that the action is on matrices instead of vectors.

H = randn(2,3)
lobs = LinearObserver(H; depth=4) # standard linear observer for matrices of size 3x4
x = randn(3,4)
lobs(x) == H*x # true

There is a right version as well:

H = randn(2,3)
lobs = LinearObserver(H, RightAction(); depth=4)
x = randn(4,3)
lobs(x) == x*H' # true

GeometricFilter.Observation — Type

Observation(observer::Observer, noise::Noise, measurement)

Encapsulation of a full observation data. This includes the actual measurement, but also how it was measured (the observer) and with how much noise. The Observation object is thus built from

an observer (AbstractObserver)

an observation noise (AbstractNoise)

a measurement, i.e., a point lying on the observer's space

GeometricFilter.ProductObserver — Method

ProductObserver(observers::AbstractObserver...)

Combine existing observers into a new one.

M = ManifoldsBase.DefaulManifold(2)
# a simple observer for the sake of this example
ob1 = IdentityObserver(M)
ob2 = IdentityObserver(M)
x = rand(M) # a point on M
ob = ProductObserver(ob1, ob2)
ob(x) # (x, x)

GeometricFilter.StochasticMotion — Method

StochasticMotion(motion::Motion, process_noise::ActionNoise)

Encapsulate the idea of a stochastic dynamical system on a manifold $\mathcal{M}$, defined by a motion $φ \colon \mathcal{M}→\mathfrak{g}$, and a noise model on the manifold.

The noise process_noise must implement get_lie_covariance_at, so must be of type AbstractActionNoise.

GeometricFilter._predict — Method

_predict(distribution::ActionDistribution, motion::Motion, process_noise::ActionNoise) :: ActionDistribution

Updates the distribution according to the motion and the process noise.

GeometricFilter.add_process_noise — Method

add_process_noise(dist::AbstractActionDistribution, noise::AbstractActionNoise)

[Deprecated]

Add process noise to a distribution of type AbstractActionDistribution.

GeometricFilter.generate_signal — Method

generate_signal(
  motions::AbstractVector{<:AbstractStochasticMotion}, x0,
  ::FilteringMode
)

Integrate one sample path the sequence of stochastic motions, starting at the point $x0$.

GeometricFilter.generate_signal — Method

generate_signal(motions::AbstractVector{<:AbstractMotion}, x0)

Integrate the sequence of motions, starting at the point $x_0$.

GeometricFilter.get_measurement — Function

get_measurement(::AbstractObserver, x) :: [point in observation manifold]

The actual observer function, mapping a point $x$ in the state space to a point in the measurement manifold.

GeometricFilter.get_motion — Function

get_motion(m::AbstractStochasticMotion) :: AbstractMotion

The underlying deterministic motion.

GeometricFilter.get_noise — Function

get_noise(m::AbstractStochasticMotion) :: AbstractActionNoise

The underlying process noise.

GeometricFilter.get_tan_observer — Function

get_tan_observer(
    ::AbstractObserver,
    A, # group action G ⊂ Diff(M)
    x, # point in M
    y=obs(x), # point in N
) :: Function

Return the pullback of the tangent map of the observer, that is the function defined by $H(ξ) = ⟨dh, ξ⋅x⟩$.

GeometricFilter.noisy_observation — Method

noisy_observation(rng::Random.AbstractRNG,
  obs::AbstractObserver,
   noise::AbstractNoise,
     state)

Observation object encapsulating a noisy version of obs(state).

GeometricFilter.observation_space — Function

observation_space(obs::AbstractObserver) :: AbstractManifold

The manifold on which the measurement lies.

GeometricFilter.observed_space — Function

observation_space(obs::AbstractObserver) :: AbstractManifold

The manifold which is observed

GeometricFilter.predict — Method

predict(distribution::AbstractActionDistribution, stochastic_motion::StochasticMotion)

Compute the update of the uncertainty (the distribution), given a motion and its associated process noise (a stochastic motion).

GeometricFilter.rigid_perturbation — Method

rigid_perturbation(rng::RNG, noise::ActionNoise, x) :: RigidMotion

Compute a random rigid motion using the covariance of the stochastic motion at $x$, where $x$ is a point on the sample space of the noise.

GeometricFilter.sensor_perturbation — Method

sensor_perturbation(rng::RNG, sm::StochasticMotion, x) :: StochasticMotion

Simulate sensor noise by adding a random rigid motion with velocity drawn from the stochastic motions noise.

GeometricFilter.simulate — Method

simulate(s::StochasticMotion, x::TM,
  ::FilteringMode,
  ) ::TM

Integrate the stochastic motion: deterministically integrate the underlying motion, and adds noise on the result.

The FilterMode can be the following:

With DataMode, the motion is integrated exactly.

With PositionPerturbation(rng), the noise is added after exact integration.

With SensorPerturbation(rng), the noise is applied to the motion, followed by an exact integration.

GeometricFilter.simulate_filter — Method

simulate_filter(
    ::ActionDistribution,
    ::Vector{<:StochasticMotion},
    ::Vector{<:Observation},
    ::FilteringMode,
      ) :: Tuple{Vector(Int), Vector(ActionDistribution)}

Runs a simulation of the filter, given a starting distribution, a vector of stochastic motions, and a vector of observations. Return a dataframe with a column containing the “streak number”, and one column containing the filtered distributions. There are three main mode to use as FilterMode:

DataMode: when it is used with real data

SensorPerturbation(rng) for simulated perturbation of sensor inputs

PositionPerturbation(rng) for simulated perturbations of positions

GeometricFilter.simulate_filter_from — Method

simulate_filter(prediction_step,
D0,
stochastic_motions,
observations)

Run a filter from stochastic motions and observations. The argument prediction_step can be used to modify the standard behaviour of predict, for instance by adding sensor noise, position noise, or both.

GeometricFilter.translate_action — Method

Translate infinitesimal action into the left action of the observation group on itself, in other words, we solve the equation $ξ⋅χ = η∘χ$ and return $η$, where $∘$ denotes the standard or dual left action of the group on itself.

GeometricFilter.update — Method

update(prior::AbstractActionDistribution, observation::AbstractObservation)

Update the uncertainty (the prior) given an observation. Typically, it is used as

update(prior, Observation(observer, noise, measurement))

with a given observer, an associated noise, and a measurement on the observation manifold (the manifold of the observer).